modified:   gui/ide.pyw 	modified:   gui/instrumentspanel.py 	modified:   gui/loopspanel.py 	modified:   helpers/coderunner.py 	modified:   helpers/datamanager.py 	modified:   helpers/instrumentsmanager.py 	modified:   lib/classes.py 	modified:   lib/smartloop.py 	modified:   server/pickle_server.py 'q' n'est pas reconnu en tant que commande interne ou externe, un programme excutable ou un fichier de commandes.

+++ lib/smartloop.py
@@ -46,9 +46,10 @@ class SmartLoop(Subject,Observer,Reloadable):
       else:
         self._value=self._value+self._step
     self._index+=1
-    if self._value>self._stop:
-      self.delete()
-      raise StopIteration
+    if self._stop!=None:
+      if self._value>self._stop:
+        self.delete()
+        raise StopIteration
     
     self.lm.updateLoop(self)
     
@@ -84,8 +85,7 @@ class SmartLoop(Subject,Observer,Reloadable):
   
     self.lm=LoopManager()
     self.lm.addLoop(self)
-  
-  
+    
   
   
   def __del__(self):
